enter.append preserves data order.
Previously, append() would insert the new element as the last child. Now, it
inserts the new element before any following sibling in the update selection.

There are two things I don't like about this implementation. First, it requires
storing a __sibling__ property on the data nodes. It would be better if this
were just temporary state in the enter selection's select method. Second, it
requires passing this sibling property to the selector a third argument. This is
non-standard and I don't like expose the data this way. However, I can't think
of another obvious way to make that state available to insert or append.

This commit also includes a minor extension to selection.insert, which is that
the before argument can now be specified as a function rather than a selector.